refactor(test): split reactive_tests.rs into modular test files
