+++ tests/reactive/mod.rs
@@ -1,0 +1,10 @@
+//! Reactive integration tests - split into modules
+
+mod async_state;
+mod batch;
+mod computed;
+mod context;
+mod effect;
+mod integration;
+mod runtime;
+mod signal;
